@@ -1,9 +1,7 @@
-#include <iostream>
+#include "cli.hpp"
+#include "config.hpp"
 
-#include "CLI.h"
-#include "config.h"
-
-int main(int argc, char const* argv[]) {
+int main(int argc, const char* argv[]) {
   CLI* cli = new CLI;
   cli->parseOpts(argc, argv);
   return cli->run();
